Add handling of special cases for phase velocity
When phase velocity c is equal to α or β in any layer, r and s vanish
and sin(0) / 0 trips up the dispersion function value. Replacing it
with 0 is a horrible idea (should have known better), replacing it
with 1 (according to it's Maclaurin series) also doesn't really fix
the issue. Now the dispersion function preforms the check each time,
and uses slightly modified values for c when a problem can appear.